Gah: Make domain shareable
to_retire can be shrouded in a mutex - that guarantees the single writer
that RCU expects anyways. refs (the current ref count) should probably
be an atomic pointer, and each Retiring generation and ReadLock should
probably just hold a pointer to it. The safety invariant we'll have to
maintain is that no ReadLock outlives the Retiring generation, which
we're betting on anyways.